♻️ refactor(behaviour)!: move click events to Text namespace
BREAKING CHANGE: ClickEvents was moved to MineJason.Text.Behaviours.Click
