+++ src/MineJason/Dialogs/StaticDialogAction.cs
@@ -1,6 +1,8 @@
 // SPDX-FileCopyrightText: (C) WithLithum & contributors 2023-2026
 // SPDX-License-Identifier: LGPL-3.0-or-later
 
+using MineJason.Text.Behaviour.Click;
+
 namespace MineJason.Dialogs;
 
 using MineJason.Events;
