+++ src/MineJason/Text/Behaviour/Click/OpenUrlClickEvent.cs
@@ -1,55 +1,54 @@
 ﻿// SPDX-FileCopyrightText: (C) WithLithum & contributors 2023-2026
 // SPDX-License-Identifier: LGPL-3.0-or-later
 
-namespace MineJason.Events;
-
 using System.Diagnostics.CodeAnalysis;
 using System.Text.Json.Serialization;
-using JetBrains.Annotations;
 using MineJason.Serialization.TextJson;
 
+namespace MineJason.Text.Behaviour.Click;
+
 /// <summary>
-/// Represents a click event that executes a command.
+/// Represents a click event that prompts to open the specified URL in the system default browser.
+/// This class cannot be inherited.
 /// </summary>
-[PublicAPI]
 [JsonConverter(typeof(ClickEventConverter))]
-public sealed class RunCommandClickEvent : ClickEvent, IEquatable<RunCommandClickEvent>
+public sealed class OpenUrlClickEvent : ClickEvent, IEquatable<OpenUrlClickEvent>
 {
     /// <summary>
-    /// Initializes a new instance of the <see cref="RunCommandClickEvent"/> class.
+    /// Initializes a new instance of the <see cref="OpenUrlClickEvent"/> class.
     /// </summary>
-    public RunCommandClickEvent()
+    public OpenUrlClickEvent()
     {
     }
-    
+
     /// <summary>
-    /// Initializes a new instance of <see cref="RunCommandClickEvent"/> with the specified value.
+    /// Initializes a new instance of <see cref="OpenUrlClickEvent"/> with the specified value.
     /// </summary>
-    /// <param name="value">The command to execute.</param>
+    /// <param name="value">The URL to open.</param>
     [SetsRequiredMembers]
-    public RunCommandClickEvent(string value)
+    public OpenUrlClickEvent(Uri value)
     {
         Value = value;
     }
     
     /// <summary>
-    /// Gets the command to execute.
+    /// Gets the URL to open.
     /// </summary>
     [JsonPropertyName("value")]
-    public required string Value { get; init; }
-    
+    public required Uri Value { get; init; }
+
     /// <inheritdoc />
-    public bool Equals(RunCommandClickEvent? other)
+    public bool Equals(OpenUrlClickEvent? other)
     {
         if (ReferenceEquals(null, other)) return false;
         if (ReferenceEquals(this, other)) return true;
-        return Value == other.Value;
+        return Value.Equals(other.Value);
     }
 
     /// <inheritdoc />
     public override bool Equals(object? obj)
     {
-        return ReferenceEquals(this, obj) || obj is RunCommandClickEvent other && Equals(other);
+        return ReferenceEquals(this, obj) || obj is OpenUrlClickEvent other && Equals(other);
     }
 
     /// <inheritdoc />
